[DEVELOPMENT ENVIRONMENT] * NEW: Image editor: Support for gamma, hue, saturation and lightness   balance. Read-only mode has been implemented. Many bug fixes everywhere.
[GB.FORM]
* NEW: Remove Brightness and Contrast properties from the ImageView 
  control.

[GB.IMAGE]
* NEW: Image.BrightnessContrast has been removed.
* NEW: New balance methods: Image.Brightness, Image.Contrast, Image.Gamma,
  Image.Hue, Image.Saturation, Image.Lightness.
* NEW: Image.BeginBalance() and Image.EndBalance() are two new methods that
  allow to group calls to the previous methods, so that all the balances
  are done in one shot.


git-svn-id: svn://localhost/gambas/trunk@5776 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/lib/image/c_color.c
@@ -62,7 +62,7 @@ void COLOR_rgb_to_hsv(int r, int g, int b, int *H, int *S, int *V)
 	{
 		f = (r == x) ? g - b : ((g == x) ? b - r : r - g);
 		i = (r == x) ? 3 : ((g == x) ? 5 : 1);
-		*H = (int)((i - (float)f / (v - x)) * 60);
+		*H = (int)((i - (double)f / (v - x)) * 60);
 		*S = ((v - x) * 255) / v;
 		*V = v;
 		if (*H == 360) *H = 0;
